backend/requirements.txt
pydantic>=2.5.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
redis>=5.0.1
celery>=5.3.6
minio>=7.2.0
boto3>=1.34.0
websockets>=12.0
httpx>=0.26.0
structlog>=24.1.0
prometheus-client>=0.19.0
sentry-sdk>=1.39.0
cryptography>=42.0.0
python-magic>=0.4.27
ffmpeg-python>=0.2.0
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.36.0
faster-whisper>=0.10.0
pyannote.audio>=3.1.0
silero-vad>=0.1.0
vllm>=0.2.0
llama-cpp-python>=0.2.0
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
pyyaml>=6.0.1
python-json-logger>=2.0.7
jsonschema>=4.20.0
email-validator>=2.1.0
pyotp>=2.9.0
qrcode>=7.4.0
pillow>=10.2.0
